requirements.txt
python-dotenv
langchain
langchain-core
langchain-community
langchain-groq
streamlit
google-generativeai
PyPDF2
faiss-cpu
huggingface_hub
transformers
langchain-huggingface  # Use dash, not capitalized
-e .
